shopper-concierge-demo/requirements_api.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-bigquery
pydantic
python-dotenv
google-generativeai
